Refactoring of frontend to workarround iOS Safari focus issues Reworked HeatingManger to use capability listeners and prefiltered device lists
