+++ package.json
@@ -8,6 +8,7 @@
     "@material-ui/core": "^3.9.0",
     "@material-ui/icons": "^3.0.2",
     "athom-api": "^2.1.156",
+    "body-scroll-lock": "^2.6.1",
     "bootstrap": "^4.2.1",
     "console-remote-client": "^0.4.4",
     "date-fns": "^2.0.0-alpha.21",
@@ -17,6 +18,7 @@
     "react-dom": "^16.7.0-alpha.0",
     "react-material-ui": "0.0.1",
     "react-router-dom": "^4.3.1",
+    "react-scroll-locky": "^1.5.0",
     "redux": "^4.0.1",
     "strongly-typed-events": "^1.4.10",
     "uuid": "^3.3.2"
@@ -50,16 +52,13 @@
   "scripts": {
     "homey:run": "npm run build:app:dev && athom -p ../homey-heating-dist app run",
     "homey:install": "npm run build:app:prod && npm run build:web:prod && athom -p ../homey-heating-dist app install",
-
     "dist": "npm run build:app:prod && npm run build:web:prod && athom -p ../homey-heating-dist app install",
-
     "build:app:dev": "webpack --mode development --config webpack.app.js",
     "build:app:prod": "webpack --mode production --config webpack.app.js",
-    
     "build:web:dev": "webpack --mode development --config webpack.react.js",
     "build:web:prod": "webpack --mode production --config webpack.react.js",
-    
-    "start:web": "webpack-dev-server --mode development --config webpack.react.js --open"
+    "start:web": "webpack-dev-server --mode development --config webpack.react.js --open --host 192.168.178.23",
+    "start:web:prod": "webpack-dev-server --mode production --config webpack.react.js --open"
   },
   "repository": {
     "type": "git",
