--- src/settings/css/site.css
@@ -1,15 +1,0 @@
-body,
-html,
-#root {
-    width: 100%;
-    height: 100%;
-}
-
-body,
-html {
-    overflow: hidden;
-}
-
-#root {
-    overflow: auto;
-}
