+++ src/settings/index.remote.html
@@ -1,0 +1,6 @@
+<html>
+    <head>
+        <meta http-equiv="refresh" content="0; url=http://192.168.178.23:8080" />
+    </head>
+    <body></body>
+</html>
